Refined Resolution Settings & Started Projectiles
The program now determines the user's screen size and sets the game resolution to that rather than simply setting it to 1920x1080.

Set window to borderless.

The health and mana rectangles on the U.I. did not load correctly on lower resolutions, they needed to be divided by the screen scale which has now been done.

A projectile class has been added along with two test models for projectiles. I will soon begin work on implementing the player's shooting mechanics.