Added Basic Player Projectiles
The player can now shoot projectiles with left click. The destination of the projectiles is loosely based off of the camera's rotation.

I'll be refining it sometime soon to make it more precise.

Otherwise the HUD has turned out fantastic! I plan to create multiple test enemy and player models next so we can work with the role swapping code later down the line.