Added Pause Menu.
Thought I'd keep it a a bit more visually minimal and keep the player able to see the game while paused, but am open to changing that if you'd want.

I plan on having pressing settings in the pause menu bring the player to the main settings screen, but they will not lose progress.(opposed to making a seperate settings screen for in game.