Updated README, DesignDoc, and minor code changes
The main focus of this update is to expand on the README and Design Document to better reflect the current state of the game but I also did a couple minor tweaks to the code.

I plan to have simple models (similar to the current model for the Cleric) for each of the main protagonists along with at least 3 enemy models done by the end of the week. After that I'll begin refinement on the player's behavior.