Character abilities & slightly refined movement
The largest change of this update is the inclusion of character abilities. Left click for a regular attack, Right click for a special attack, or press Q or E for a special ability. I added a couple of test abilities but these are subject to change.

The player can now move in four directions and they can sprint. Upon reaching the boss area they can move in a more wide open range.

The test level has been slightly altered to match more of what I envision the first level to look like, though I still plan to work with it a lot more by texturing it and creating some background models to make it look more interesting.

Overall I plan to refine the player's camera and shooting right after this, I hope to have that finished by 3:00 PM tomorrow.