Many Changes to player and projectile
New class "AttackPattern" has been added to handle any and all bullet patterns. Feel free to add your own custom bullet patterns.

Now, when you want something to shoot a projectile, all you need to provide AttackPattern with is:
the function (attack pattern) you wish to use
the position where the projectile is coming from (EX: player.Pos, enemy.Pos)
where the projectile is going (EX: camera.LookAt, player.Pos)
the speed of the projectile (0 for melee, will be implemented later)
and the projectile list that will hold all of these projectiles.

If you are confused as to how it works, take a look at the player classes "Abilities" function and locate any point in which attack pattern is used.
The player class has been updated to accommodate to the new projectile changes but also has been altered to better prepare for character switching (the player's stats can now be changed depending on the character selected)

I still intend on updating the camera class as well so that the player can have a bit more movement with the camera but all of that only affects the player class. I mostly wanted to push this update out now as it will help Hayden develop the enemy class.