Added basic melee attacks
The projectile class has been renamed to the attack class as it now hosts the ability to control both melee and ranged features.
If you wish for something to use a melee attack simply pass it the singleMel function from the AttackPattern class.

The player's SwitchRole debugging function has been updated so that it changes more than just the model itself and actually shifts all the player's stats to the chosen character (it seems to have a strange effect on jumping currently but otherwise works fine)