Added Level Select Screen.
Currently it is just superficial and the text is moreso a placeholder. If we get to adding a second level, the code is in place to start working on that, but I figure we'll need to implement level files that we don't have yet.

Added a 2d projectile image. feel free to remove if you'd prefer keep this aspect 3D, I wanted to get down 2d sprites in 3d space a bit better, so I made the class to implement this and thought that the projectiles was a good place to implement