Imported some basic models for the other three main characters
I've included basic models for all of the possible characters. I'm currently in the middle of working on some more complex ones with animations and such but those won't be done for a while.
My next plan is to make some simple enemy models and then I'll be focusing on improving the gameplay.