Created a Level Complete Scene.
This also has code implemented to make the scoreboard function. If you want to clear out the high schores, just empty the txt file.

The scorebaord doesn't need much more implementing to work witrh the code I have for the level complete screen now, so I'll work on that this week as well and hopefully have time to do the character Icons.