Sound Design and Slight Gameplay Refinements
Since it's likely that we will not be hearing from Lucas before the end of the day I went ahead and handled the sound design. All royalty free. it has been added to the credits.

I also made it so that the audio settings affect the sounds.

Otherwise I gave every hero a second attack and two special abilities.